bugfix/lighthouse-report-images-a11y-shift: refactoring; fix large layout shifts
