+++ models/index.ts
@@ -3,3 +3,5 @@ export type { PricingPlanItem, PlanType } from './pricing-plan.model';
 export type { CaseStudy, ProjectFeedback } from './case-study.model';
 
 export type { Service } from './service.model';
+
+export type { ReactChildren } from './common';
